+++ NEWS.md
@@ -1,8 +1,9 @@
-# phenoptr 0.3.1.9000
+# phenoptr 0.3.2
+**2022-01-03**
 
-- Correctly parse phenotype expressions containing slashes (#18)
 - Fix `parse_composite_info` to work with older composite images
   that are missing ImageDescription (akoyabio/phenoptrReports#52)
+- Correctly parse phenotype expressions containing slashes (#18)
 - Tighten up the column exclusion criteria for 
   `read_cell_seg_data` with `col_select="phenoptrReports"` (#16)
 - Update tests to work with testthat 3.1.